+++ src/objs/observers/json_logger.cpp
