+++ src/objs/observers/json_logger.h
