+++ src/objs/observers/live_terminal.cpp
