+++ tests/terminal_output.cpp
@@ -1,4 +1,4 @@
-#include <test/assert.h>
+#include <test.h>
 #include <thread>
 
 using namespace std;
